refactor: 예외 클래스 패키지 수정
