+++ src/main/java/racingcar/exception/carname/CarNameException.java
@@ -1,4 +1,4 @@
-package racingcar.validator.exception.carname;
+package racingcar.exception.carname;
 
 public class CarNameException extends RuntimeException{
     public CarNameException(String message) {
