+++ src/main/java/racingcar/exception/carname/CarNameLengthOverException.java
@@ -1,4 +1,4 @@
-package racingcar.validator.exception.carname;
+package racingcar.exception.carname;
 
 public class CarNameLengthOverException extends CarNameException {
     public CarNameLengthOverException(int length) {
