+++ src/main/java/racingcar/exception/carname/CarNameNotExistsException.java
@@ -1,4 +1,4 @@
-package racingcar.validator.exception.carname;
+package racingcar.exception.carname;
 
 public class CarNameNotExistsException extends CarNameException {
     public CarNameNotExistsException() {
