+++ src/main/java/racingcar/exception/carname/CarNameNullPointerException.java
@@ -1,4 +1,4 @@
-package racingcar.validator.exception.carname;
+package racingcar.exception.carname;
 
 public class CarNameNullPointerException extends CarNameException {
     public CarNameNullPointerException() {
