+++ src/main/java/racingcar/exception/trialcount/TrialCountException.java
@@ -1,4 +1,4 @@
-package racingcar.validator.exception;
+package racingcar.exception.trialcount;
 
 public class TrialCountException extends RuntimeException {
     public TrialCountException(String message) {
