+++ src/main/java/racingcar/exception/trialcount/TrialCountNotPositiveException.java
@@ -1,4 +1,4 @@
-package racingcar.validator.exception;
+package racingcar.exception.trialcount;
 
 public class TrialCountNotPositiveException extends TrialCountException {
     public TrialCountNotPositiveException() {
