+++ src/test/java/racingcar/TrialCountTest.java
@@ -8,10 +8,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.ValueSource;
 import racingcar.domain.TrialCount;
-import racingcar.util.StringUtil;
-import racingcar.validator.exception.TrialCountException;
+import racingcar.exception.trialcount.TrialCountException;
 
 public class TrialCountTest {
